added fundamentals to learning resources
