+++ Algo-Learning-Resources/Fundamentals/Linked-List-Fundamentals/Linked-List-Patterns.md
